Fiz o instalador do MoneroD

+++ projects/api/downloadinstall.py
@@ -9,7 +9,7 @@ sys.path.append(CURRENTDIR);
 from api.process import Process;
 
 class DownloadInstall():
-    def __init__(self, file, url):
+    def __init__(self, file, url, log=None):
         self.url = url;
         self.file = file;
         self.tmp = "/tmp/" + str(uuid.uuid4());
@@ -30,27 +30,32 @@ class DownloadInstall():
             else:
                 os.chmod(path, st.st_mode | stat.S_IRGRP | stat.S_IROTH);
 
-    def download(self):
-        if os.path.exists("/tmp/" + self.file):
+    def download(self, force=True):
+        if os.path.exists("/tmp/" + self.file) and force == True:
             os.unlink("/tmp/" + self.file);
-        process = Process("wget "+ self.url +" -O /tmp/" + self.file);
-        process.run();
+        if not os.path.exists("/tmp/" + self.file):
+            process = Process("wget "+ self.url +" -O /tmp/" + self.file);
+            process.run();
         if self.file.find("tar.gz") > 0:
             process = Process("tar xzvf /tmp/" + self.file + " -C "+ self.tmp +" --strip-components 1");
             process.run();
         elif self.file.find(".zip") > 0:
             process = Process("unzip /tmp/" + self.file + " -d "+ self.tmp +"");
             process.run();
+        elif self.file.find("tar.bz2") > 0:
+            process = Process("tar -xjvf /tmp/"+ self.file +" -C "+ self.tmp  +" --strip-components 1");
+            process.run();
         else:
-            print(self.file);
             process = Process("tar --strip-components 1 -C "+ self.tmp +" -xf /tmp/" + self.file);
             process.run();
+        return self.tmp;
 
-    def extract(self, path):
+    def extract(self, path, permission=True):
         if not os.path.exists(path):
             os.makedirs(path, exist_ok=True);
         shutil.copytree(self.tmp, path , dirs_exist_ok=True)
-        self.__permission__(path);
+        if permission == True:
+            self.__permission__(path);
     
     def make(self):
         self.download();
